Move FoalModule to interfaces/
