+++ packages/core/src/index.ts
@@ -3,5 +3,5 @@ export * from './interfaces';
 export { combineHooks } from './utils';
 
 export * from './errors';
-export { Foal, FoalModule } from './foal';
+export { Foal } from './foal';
 export { ServiceManager, Service } from './service-manager';
