+++ packages/core/src/interfaces/index.ts
@@ -2,3 +2,4 @@ export * from './contexts';
 export * from './controller-and-routes';
 export * from './middlewares';
 export * from './utils';
+export * from './module';
