+++ packages/core/src/interfaces/module.ts
@@ -1,0 +1,9 @@
+import { Controller } from './controller-and-routes';
+import { Hook, Type } from './utils';
+
+export interface FoalModule {
+  services: Type<any>[];
+  controllers?: Controller[];
+  hooks?: Hook[];
+  modules?: { module: FoalModule, path?: string }[];
+}
